feat: payment with update server side code
