+++ client/src/components/CheckoutForm.jsx
@@ -50,7 +50,7 @@ export default function CheckoutForm() {
     const { error } = await stripe.confirmPayment({
       elements,
       confirmParams: {
-        return_url: `http://localhost:5173/rest?`},
+        return_url: `http://localhost:5173/payment-success?`},
     });
 
     if (error.type === 'card_error' || error.type === 'validation_error') {
